+++ examples/scripts/DFN.py
@@ -20,7 +20,7 @@ param.process_geometry(geometry)
 
 # set mesh
 var = pybamm.standard_spatial_vars
-var_pts = {var.x_n: 75, var.x_s: 21, var.x_p: 55, var.r_n: 50, var.r_p: 50}
+var_pts = {var.x_n: 30, var.x_s: 30, var.x_p: 30, var.r_n: 10, var.r_p: 10}
 mesh = pybamm.Mesh(geometry, model.default_submesh_types, var_pts)
 
 # discretise model
@@ -29,8 +29,9 @@ disc.process_model(model)
 
 # solve model
 t_eval = np.linspace(0, 0.2, 100)
-# solver = model.default_solver
-solver = pybamm.KLU(rtol=1e-6, atol=1e-6)
+solver = model.default_solver
+solver.rtol = 1e-3
+solver.atol = 1e-6
 solution = solver.solve(model, t_eval)
 
 # plot