+++ results/2019_08_sulzer_thesis/lead_acid_discharge.py
@@ -131,9 +131,7 @@ def discharge_times_and_errors(compute):
         except FileNotFoundError:
             models_times_and_voltages = pybamm.get_infinite_nested_dict()
         models = [
-            pybamm.lead_acid.Full(
-                {"surface form": "algebraic"}, name="Full"
-            ),
+            pybamm.lead_acid.Full({"surface form": "algebraic"}, name="Full"),
             pybamm.lead_acid.LOQS(name="LOQS"),
             # pybamm.lead_acid.FOQS(name="FOQS"),
             # pybamm.lead_acid.Composite(name="Composite"),