+++ results/2019_xx_2plus1D_pouch/pouch_cell_model.py
@@ -17,7 +17,7 @@ options = {
     "dimensionality": 2,
     "thermal": "x-lumped",
 }
-model = pybamm.lithium_ion.SPM(options)
+model = pybamm.lithium_ion.DFN(options)
 geometry = model.default_geometry
 
 # load parameters and process model and geometry
@@ -31,11 +31,11 @@ var = pybamm.standard_spatial_vars
 submesh_types = model.default_submesh_types
 
 # cube root sequence in particles
-r_n_edges = np.linspace(0, 1, 11) ** (1 / 3)
+r_n_edges = np.linspace(0, 1, 5) ** (1 / 3)
 submesh_types["negative particle"] = pybamm.MeshGenerator(
     pybamm.UserSupplied1DSubMesh, submesh_params={"edges": r_n_edges}
 )
-r_p_edges = np.linspace(0, 1, 11) ** (1 / 3)
+r_p_edges = np.linspace(0, 1, 5) ** (1 / 3)
 submesh_types["positive particle"] = pybamm.MeshGenerator(
     pybamm.UserSupplied1DSubMesh, submesh_params={"edges": r_p_edges}
 )
@@ -60,7 +60,7 @@ y4 = np.linspace(centre_tab_p + l_tab_p / 2, l_y, 3)  # mesh from pos tab to cel
 y_edges = np.concatenate((y0, y1[1:], y2[1:], y3[1:], y4[1:]))
 
 # cube root sequence in z direction
-z_edges = np.linspace(0, 1, 10) ** (1 / 3)
+z_edges = np.linspace(0, 1, 5) ** (1 / 3)
 submesh_types["current collector"] = pybamm.MeshGenerator(
     pybamm.UserSupplied2DSubMesh,
     submesh_params={"y_edges": y_edges, "z_edges": z_edges},
@@ -79,7 +79,7 @@ mesh = pybamm.Mesh(geometry, submesh_types, var_pts)
 
 # discretise model
 disc = pybamm.Discretisation(mesh, model.default_spatial_methods)
-disc.process_model(model)
+disc.process_model(model, check_model=False)
 
 # discharge timescale
 tau = param.evaluate(pybamm.standard_parameters_lithium_ion.tau_discharge)
@@ -88,7 +88,6 @@ tau = param.evaluate(pybamm.standard_parameters_lithium_ion.tau_discharge)
 t_end = 900 / tau
 t_eval = np.linspace(0, t_end, 120)
 solver = pybamm.IDAKLUSolver(atol=1e-6, rtol=1e-6, root_tol=1e-6)
-# solver = pybamm.CasadiSolver(atol=1e-6, rtol=1e-6, root_tol=1e-6)
 solution = solver.solve(model, t_eval)
 
 # TO DO: 2+1D automated plotting